requirements.txt
streamlit==1.28.0
openai>=1.3.0 
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.0.0
pydub==0.25.1
gtts==2.4.0
pytesseract==0.3.10
pytest==7.4.3